添加了显示范围调整,完善了获取笔记列表方法,Query类支持了in 查询的情况,适当调整了log类

+++ app/models/Items.php
@@ -1,25 +1,22 @@
 <?php
+
 namespace n\models;
 
 use nxn\db\ActiveRecord;
 use nxn\db\Query;
 
-class Items extends ActiveRecord {
-    /**
-     * 表示全局可见，及在任何一个目录下都可以看到
-     */
+class Items extends ActiveRecord
+{
+
+    //表示全局可见，及在任何一个目录下都可以看到
     const SHOW_GLOBAL = 20;
+    const SHOW_LOCAL = 10;
+    const STATUS_ENABLE = 10;
+    // 暂存的状态
+    const STATUS_STATING = 20;
+    //归档
+    const STATUS_DRAFT = 30;
 
-    /**
-     *  启用
-     */
-
-    const ENABLE = 2;
-
-    /**
-     * 放入草稿箱
-     */
-    const DRAFT = 3;
 
     public static function tableName()
     {
@@ -58,25 +55,29 @@ class Items extends ActiveRecord {
      * Description:
      *
      */
-    public static function getItems($fid, $status = [])
+    public static function getItems($fid, $userId, array $status = [])
     {
-        $userId = $_SESSION['user_id'];
-        $statusPart = '';
-        if($status){
-            if(count($status) === 1){
-                $statusPart = ' AND `status` = :status ';
-            }else{
-                $statusPart = ' AND `status` in (:status) ';
-            }
-        }
-        $sql = 'SELECT * FROM `items` WHERE `user_id` = :user_id ' . $statusPart.  ' AND (`visible_range` = :show_global OR `fid` = :fid )  ORDER BY `visible_range` DESC, `rank` DESC';
-        $param = [':show_global' => self::SHOW_GLOBAL, ':fid' => $fid, ':user_id' => $userId, ':status' => join(',', $status)];
-        return Query::all($sql, $param);
+        // 默认查出启用的状态
+        $status = array_unique(array_merge([self::STATUS_ENABLE], $status));
+        $sql = 'SELECT *
+FROM `items`
+WHERE `user_id` = :user_id AND `status` IN (:status) AND (`visible_range` = :show_global OR `fid` = :fid)
+ORDER BY `visible_range` DESC, `rank` DESC';
+        $params = [
+            ':show_global' => self::SHOW_GLOBAL,
+            ':fid' => (int)$fid,
+            ':user_id' => (int)$userId,
+            ':status' => $status,
+        ];
+        return Query::all($sql, $params);
     }
 
-    public static function deleteItem($id)
+    public static function deleteItem($id, $userId)
     {
         $item = self::load($_REQUEST['id']);
+        if ($item->user_id != $userId) {
+            throw new \Exception('items 不属于该用户', 402);
+        }
         $status = $item->delete();
         return $status;
     }
@@ -89,14 +90,14 @@ class Items extends ActiveRecord {
 
     public function draft()
     {
-        $this->status = Items::DRAFT;
+        $this->status = Items::STATUS_DRAFT;
         $updated = $this->save(false);
         return $updated;
     }
 
     public function toggleStatus()
     {
-        $this->status = $this->status == self::DRAFT ? self::ENABLE : Items::DRAFT;
+        $this->status = $this->status == self::STATUS_DRAFT ? self::STATUS_ENABLE : Items::STATUS_DRAFT;
         $updated = $this->save(false);
         return $updated;
     }
@@ -137,6 +138,20 @@ class Items extends ActiveRecord {
         }
     }
 
+    public static function toggleVisibleRange($id, $userId)
+    {
+        $item = self::load($id);
+        if (empty($item)) {
+            throw new \Exception('items 无效', 402);
+        }
+        if ($item->user_id != $userId) {
+            throw new \Exception('items 不属于该用户', 402);
+        }
+        $item->visible_range = $item->visible_range == self::SHOW_GLOBAL ? self::SHOW_LOCAL : self::SHOW_GLOBAL;
+        $item->save();
+        return $item->visible_range;
+    }
+
     /**
      * beforeInsert.
      * @access
@@ -147,11 +162,13 @@ class Items extends ActiveRecord {
      * Time: xx:xx
      * Description:
      */
-   public function beforeInsert()
-   {
-       $this->user_id = $_SESSION['user_id'];
-       $sql = "select rank from items where user_id = :user_id ORDER BY rank DESC";
-       $maxRank = Query::scalar($sql, [':user_id' => $this->user_id]);
-       $this->rank = $maxRank + 2;
-   }
+    public function beforeInsert()
+    {
+        $this->user_id = $_SESSION['user_id'];
+        $sql = 'select rank from items where user_id = :user_id ORDER BY rank DESC';
+        $maxRank = Query::scalar($sql, [':user_id' => $this->user_id]);
+        $this->rank = $maxRank + 2;
+    }
+
+
 }
